Refactored context
Made combine reducers work (although was very rude, used "as").
Changed state structure in order for it to fit the reducers.
Going too flat in state structure proved itself unnecessary.
I intend to use "users" state as a database wannabe.
As far as I can tell: a little bit of clean up and we are good to
move forward.